refactor(infrastructure): introduce @WorkspaceScope meta-annotation
Replace the repeated
`@Scope(value = WorkspaceScope.SCOPE_NAME, proxyMode = ScopedProxyMode.TARGET_CLASS)`
boilerplate (used on ~30 beans) with a composed `@WorkspaceScope`
meta-annotation. The proxy mode is exposed via @AliasFor (default
TARGET_CLASS), so the few interface-proxied @Bean executors use
`@WorkspaceScope(proxyMode = ScopedProxyMode.INTERFACES)`.

The Spring Scope SPI implementation is renamed WorkspaceScope ->
WorkspaceBeanScope to free the `WorkspaceScope` name for the annotation;
SCOPE_NAME stays on the implementation and is referenced by the meta-annotation.
