+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/references/model/AnnotationRepository.java
@@ -23,8 +23,6 @@ package com.github._1c_syntax.bsl.languageserver.references.model;
 
 import com.github._1c_syntax.bsl.languageserver.context.symbol.AnnotationSymbol;
 import com.github._1c_syntax.bsl.languageserver.infrastructure.WorkspaceScope;
-import org.springframework.context.annotation.Scope;
-import org.springframework.context.annotation.ScopedProxyMode;
 import org.springframework.stereotype.Component;
 
 import java.net.URI;
@@ -40,7 +38,7 @@ import java.util.concurrent.ConcurrentHashMap;
  * регистре, а поиск ведётся регистронезависимо.
  */
 @Component
-@Scope(value = WorkspaceScope.SCOPE_NAME, proxyMode = ScopedProxyMode.TARGET_CLASS)
+@WorkspaceScope
 public class AnnotationRepository {
 
   private final Map<String, AnnotationSymbol> annotations = new ConcurrentHashMap<>();
