+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/references/model/LocationRepository.java
@@ -22,8 +22,6 @@
 package com.github._1c_syntax.bsl.languageserver.references.model;
 
 import com.github._1c_syntax.bsl.languageserver.infrastructure.WorkspaceScope;
-import org.springframework.context.annotation.Scope;
-import org.springframework.context.annotation.ScopedProxyMode;
 import org.springframework.stereotype.Component;
 
 import java.net.URI;
@@ -37,7 +35,7 @@ import java.util.stream.Stream;
  * Хранилище расположений обращений к символам.
  */
 @Component
-@Scope(value = WorkspaceScope.SCOPE_NAME, proxyMode = ScopedProxyMode.TARGET_CLASS)
+@WorkspaceScope
 public class LocationRepository {
   /**
    * Список обращений к символу, сгруппированный по URI.
