+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/types/DereferenceMemberMatcher.java
@@ -45,8 +45,6 @@ import org.antlr.v4.runtime.tree.TerminalNode;
 import org.eclipse.lsp4j.Position;
 import org.eclipse.lsp4j.Range;
 import org.jspecify.annotations.Nullable;
-import org.springframework.context.annotation.Scope;
-import org.springframework.context.annotation.ScopedProxyMode;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
@@ -60,7 +58,7 @@ import java.util.Optional;
  * чтобы фасад типов оставался в рамках одного класса.
  */
 @Component
-@Scope(value = WorkspaceScope.SCOPE_NAME, proxyMode = ScopedProxyMode.TARGET_CLASS)
+@WorkspaceScope
 @RequiredArgsConstructor
 public class DereferenceMemberMatcher {
