+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/types/index/SymbolTypeIndex.java
@@ -37,8 +37,6 @@ import com.github._1c_syntax.bsl.parser.description.CollectionTypeDescription;
 import com.github._1c_syntax.bsl.parser.description.TypeDescription;
 import lombok.RequiredArgsConstructor;
 import org.jspecify.annotations.Nullable;
-import org.springframework.context.annotation.Scope;
-import org.springframework.context.annotation.ScopedProxyMode;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
 
@@ -62,7 +60,7 @@ import java.util.concurrent.ConcurrentHashMap;
  * {@code ExpressionTypeInferencer}'ом по выражению инициализации.
  */
 @Component
-@Scope(value = WorkspaceScope.SCOPE_NAME, proxyMode = ScopedProxyMode.TARGET_CLASS)
+@WorkspaceScope
 @RequiredArgsConstructor
 public class SymbolTypeIndex {
