+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/types/scope/GlobalSymbolScope.java
@@ -24,8 +24,6 @@ package com.github._1c_syntax.bsl.languageserver.types.scope;
 import com.github._1c_syntax.bsl.languageserver.context.symbol.Symbol;
 import com.github._1c_syntax.bsl.languageserver.infrastructure.WorkspaceScope;
 import org.jspecify.annotations.Nullable;
-import org.springframework.context.annotation.Scope;
-import org.springframework.context.annotation.ScopedProxyMode;
 import org.springframework.stereotype.Component;
 
 import java.util.Collection;
@@ -62,7 +60,7 @@ import java.util.concurrent.ConcurrentHashMap;
  * единой Symbol-точкой входа.
  */
 @Component
-@Scope(value = WorkspaceScope.SCOPE_NAME, proxyMode = ScopedProxyMode.TARGET_CLASS)
+@WorkspaceScope
 public class GlobalSymbolScope {
 
   /**
